Anonymous AMD Support
This is a second attempt at #5512.

It adds an option named `amdModule` to app.import. It causes any anonymous AMD module in the imported file to be registered under the given name.

    app.import('bower_components/some-dep/some-dep.js', { amdModule: 'some-dep' });

This should work correctly with a large number of libraries that:

 - check for `typeof define === 'function' && define.amd`
 - call `define` with just a factory function, or a dependency list and factory function.

PR #5512 was reverted because the `stew.map` was running over every file in bower and vendor, which was prohibitvely expensive, and which caused binary files to get rewritten as ascii. To address those concerns, I switched to using Funnel to pull out only the explicitly named amdModules into a separate tree first.